% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App.R
\docType{data}
\name{path}
\alias{path}
\title{TODO: source correct the directory path
Set your own path to your working directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
path
}
\description{
TODO: source correct the directory path
Set your own path to your working directory
}
\keyword{datasets}
